+++ math/coordinate.hpp
@@ -28,7 +28,7 @@ public:
      * @param o The origin of the coordinate system
      * @param wCoord The homogeneous coordinate
      */
-    Coordinate(Vector3 u, Vector3 v, Vector3 w, Vector3 o, float wCoord) {
+    Coordinate(Vector3 u, Vector3 v, Vector3 w, Vector3 o, double wCoord) {
 
         matrix[0][0] = u.x;
         matrix[1][0] = u.y;
@@ -99,7 +99,7 @@ Coordinate translation(const Coordinate coord, const Vector3 v);
  * @param theta The rotation angle, in radians 
  * @return Coordinate 
  */
-Coordinate rotationX(const Coordinate coord, float theta);
+Coordinate rotationX(const Coordinate coord, double theta);
 
 /**
  * @brief Rotate the coordinate system by theta radians on the Y axis.
@@ -108,7 +108,7 @@ Coordinate rotationX(const Coordinate coord, float theta);
  * @param theta The rotation angle, in radians 
  * @return Coordinate 
  */
-Coordinate rotationY(const Coordinate coord, float theta);
+Coordinate rotationY(const Coordinate coord, double theta);
 
 /**
  * @brief Rotate the coordinate system by theta radians on the Z axis.
@@ -117,7 +117,7 @@ Coordinate rotationY(const Coordinate coord, float theta);
  * @param theta The rotation angle, in radians 
  * @return Coordinate 
  */
-Coordinate rotationZ(const Coordinate coord, float theta);
+Coordinate rotationZ(const Coordinate coord, double theta);
 
 /**
  * @brief Scale the coordinate system by a scale vector
