+++ math/vector3.cpp
@@ -1,7 +1,7 @@
 #include "vector3.hpp"
 #include <cmath>
 
-float dot(const Vector3 u, const Vector3 v) {
+double dot(const Vector3 u, const Vector3 v) {
     return u.x * v.x + u.y * v.y + u.z * v.z; 
 }
 
@@ -9,7 +9,7 @@ Vector3 cross(const Vector3 u, const Vector3 v) {
     return Vector3( u.y*v.z - u.z*v.y, u.z*v.x - u.x*v.z, u.x*v.y - u.y*v.x);
 }
 
-float Vector3::modulus() const {
+double Vector3::modulus() const {
     return sqrt(x * x + y * y + z * z);
 }
